requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --index-url=https://mirrors.aliyun.com/pypi/simple/ --no-emit-index-url --output-file=requirements.txt requirements/production.in
#
aerich==0.4.3
    # via -r requirements/production.in
aiosqlite==0.16.0
    # via tortoise-orm
asyncpg==0.21.0
    # via -r requirements/production.in
certifi==2020.12.5
    # via
    #   httpx
    #   requests
    #   sentry-sdk
chardet==4.0.0
    # via requests
click==7.1.2
    # via
    #   aerich
    #   typer
    #   uvicorn
ddlparse==1.9.0
    # via aerich
fastapi==0.63.0
    # via -r requirements/production.in
gunicorn==20.0.4
    # via -r requirements/production.in
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.12.2
    # via httpx
httptools==0.1.1
    # via uvicorn
httpx==0.16.1
    # via -r requirements/production.in
idna==2.10
    # via
    #   requests
    #   rfc3986
iso8601==0.1.13
    # via tortoise-orm
pydantic==1.7.3
    # via
    #   -r requirements/production.in
    #   aerich
    #   fastapi
pyparsing==2.4.7
    # via ddlparse
pypika==0.44.1
    # via tortoise-orm
python-dotenv==0.15.0
    # via uvicorn
python-multipart==0.0.5
    # via -r requirements/production.in
pytz==2020.5
    # via tortoise-orm
pyyaml==5.3.1
    # via uvicorn
requests==2.25.1
    # via -r requirements/production.in
rfc3986[idna2008]==1.4.0
    # via httpx
sentry-sdk==0.19.5
    # via -r requirements/production.in
six==1.15.0
    # via python-multipart
sniffio==1.2.0
    # via
    #   httpcore
    #   httpx
starlette==0.13.6
    # via fastapi
tortoise-orm==0.16.19
    # via
    #   -r requirements/production.in
    #   aerich
typer==0.3.2
    # via -r requirements/production.in
typing-extensions==3.7.4.3
    # via aiosqlite
urllib3==1.26.2
    # via
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.13.3
    # via -r requirements/production.in
uvloop==0.14.0
    # via uvicorn
watchgod==0.6
    # via uvicorn
websockets==8.1
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
